refactor: consolidate CPU timing behind shared include/hip/timing.h
Eliminates duplicated timing infrastructure across the repo:

- Create include/hip/timing.h with three inline free functions:
  hipdnn_ep_timing_enabled(), record_elapsed(), elapsed_since().
  Pure C++ stdlib (<chrono>, <cstdlib>), no LLVM or HIP deps.

- Remove 3 identical local copies of record_elapsed/elapsed_since
  from hipdnn_ep_runtime_state.cpp, InferenceState.cpp, pass_main.cpp.

- Remove 3 identical getenv("HIPDNN_EP_TIMING") lambdas from the
  same files; all now call hipdnn_ep_timing_enabled().

- Update include/hip/debug_log.h to delegate hipdnn_ep_timing_enabled()
  to timing.h via #include "timing.h".

- Replace raw std::chrono::duration<double> inside CompilerDriver.cpp
  logPhase lambda and OnnxToHip.cpp logSubpass lambda with
  record_elapsed()/elapsed_since().

- Replace raw chrono one-shot measurements in hip-onnx-runner.cpp
  with elapsed_since().

- Add include/ to CMake include paths for runtime bitcode compilation,
  MLIR backend plugins, and hip-onnx-runner.

Free functions returning double seconds are output-agnostic -- callers
choose their own sink (fprintf, llvm::errs, std::cout). This keeps
measurement separate from output, which avoids coupling timing.h to
any particular logging framework.

Made-with: Cursor
