+++ backend-mlir-compiler/level-1-pass/CMakeLists.txt
@@ -29,6 +29,7 @@ target_link_libraries(${LIB_NAME} PRIVATE
 target_include_directories(${LIB_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/src
     ${CMAKE_CURRENT_BINARY_DIR}/..  # For generated proto headers from parent
+    ${CMAKE_SOURCE_DIR}/include     # Shared project headers (timing.h, etc.)
 )
 
 target_compile_definitions(${LIB_NAME} PRIVATE
